Add Makefile for backdoor specific tasks
The Makefile is aimed towards the Vagrant machine specified by the root
Vagrantfile.

Add an empty/basic Dockerfile for compilation testing.

The `AUTO_GOPATH=1` addition to the root Makefile is to hotfix a Go
installation issue within the Vagrant VM. The issue is just that the
environmental variable `$GO_PATH` is not found.